requirements.txt
ctransformers
faiss-cpu
fastapi
ipykernel
langchain
pypdf
python-box
python-dotenv
sentence-transformers
uvicorn